Merge changes of a stash pop
I'm actually not fully sure what went on here. I
tried to pop my stashed changes and that is when I
noticed the head was detached and then I created
and new branch for the changes I made while in the
detached head state but still needed to merge/
figure out the stuff that I had popped.